simple, report: change heading appearance
For 'simple', there is now an option to number section headings. If a numbering
format is provided, all of the headings will be the same size (12pt). If not,
it will use the default of larger sizes for higher headings.

For 'report', headings are all now the same size (14pt), since numbering is
always on.